added package Provider & created classes for Quiz
