+++ lib/Template/questionItem.dart
@@ -1,0 +1,9 @@
+class QuestionItem {
+  final String category;
+  final String question;
+  final String correctAnswer;
+  final List<String> incorrectAnswer;
+
+  QuestionItem(
+      this.category, this.question, this.correctAnswer, this.incorrectAnswer);
+}
