+++ pubspec.yaml
@@ -29,7 +29,10 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.0
   custom_navigator: ^0.3.0
-
+  provider: any
+  http: any
+  
+dev_dependencies:  
   flutter_test:
     sdk: flutter
